pa0.6.8 - Reworks project's folders structure
The following list shows the changes in the folders structure of the
project:

     index.html -> app/index.html
     info.mjs -> app/info.mjs
     manifest.json -> app/manifest.json
     serv.py -> app/slave.py
     static/css/cells.css -> app/static/css/cells.css
     static/css/main.css -> app/static/css/main.css
     static/css/panel.css -> app/static/css/panel.css
     static/css/table.css -> app/static/css/table.css
     static/icon.png -> app/static/icon.png
     static/js/temp.js -> app/static/js/temp.js
     static/js/test_data.js -> app/static/js/test_data.js
     static/yicon.png -> app/static/yicon.png
     asdf/index.html -> test/index.html

This was mainly due to the implementation of the backend in `python`,
and a new `run.sh` command was added to ease launching the app for
testing purposes.

The whole restructure maintains compatibility with `vite`, so one can
make use of it while developing and use `bottle` as live server.